new update: creation of community technology

+++ src/main/java/com/connect/system/domain/model/Account/AccountInformation/Location.java
@@ -14,7 +14,7 @@ public class Location {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id_location;
+    private Integer id_location;
 
     private String country;
     private String state;
