+++ src/main/java/com/connect/system/domain/model/System/Squad/MemberDashboardSquad.java
@@ -13,7 +13,7 @@ public class MemberDashboardSquad {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id_dashboardMembers;
+    private int id_dashboardMembers;
 
     @JsonIgnore
     @OneToOne
