+++ src/main/java/com/connect/system/domain/repository/System/DashboardMembersSquadRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DashboardMembersSquadRepository extends JpaRepository<MemberDashboardSquad, Long> {
+public interface DashboardMembersSquadRepository extends JpaRepository<MemberDashboardSquad, Integer> {
 }
