+++ src/main/java/com/connect/system/domain/repository/System/SquadRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SquadRepository extends JpaRepository<Squad, Long> {
+public interface SquadRepository extends JpaRepository<Squad, Integer> {
     @Query("SELECT DISTINCT s FROM Squad s LEFT JOIN FETCH s.allMembers.members WHERE s.id_squad = :id_squad")
-    Squad findSquadWithMembers(@Param("id_squad") Long id_squad);
+    Squad findSquadWithMembers(@Param("id_squad") Integer id_squad);
 }
