+++ src/main/java/com/connect/system/domain/repository/User/AcademicEducationRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AcademicEducationRepository extends JpaRepository<AcademicEducation, Long> {
+public interface AcademicEducationRepository extends JpaRepository<AcademicEducation, Integer> {
     @Query("SELECT c FROM AcademicEducation c WHERE c.id_dashboardStudies = :id_dashboardStudies")
-    List<AcademicEducation> findDashboardById(@Param("id_dashboardStudies") Long dashboard_id);
+    List<AcademicEducation> findDashboardById(@Param("id_dashboardStudies") Integer dashboard_id);
 }
