+++ src/main/java/com/connect/system/domain/repository/User/DashboardStudiesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DashboardStudiesRepository extends JpaRepository<DashboardStudies, Long> {
+public interface DashboardStudiesRepository extends JpaRepository<DashboardStudies, Integer> {
 }
