+++ src/main/java/com/connect/system/service/Impl/PasswordServiceImpl.java
@@ -46,7 +46,7 @@ public class PasswordServiceImpl implements PasswordService {
         }
     }
 
-    public Person toUpdatePassword(Person updatedNewPassword, PasswordDTO passwordDTO, Long id) {
+    public Person toUpdatePassword(Person updatedNewPassword, PasswordDTO passwordDTO, Integer id) {
         Person authenticatedUser = (Person) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
         if (!authenticatedUser.getId().equals(id)) {
